requirements.txt
openrouter
requests
python-dotenv
Pillow
